pyproject.toml
[tool.poetry]
name = "bot-telega-nikolaj"
version = "0.1.0"
description = ""
authors = ["theRacoonPrime <stoletov314@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.4"
python-dotenv = "^1.0.1"
urllib3 = "^2.2.3"
update = "^0.0.1"
aiogram = "^3.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

